Article update

+++ src/persistence/PersonManager.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Local
 public interface PersonManager extends Serializable {
     public String addPerson(Person person, String password);
-    public Person getUserByID(int id);
+    public Person getUserByID(Long id);
     public Person getUserByCreditentials(String user,String password);
 }
